+++ server/pages/api/one.json.js
@@ -1,0 +1,5 @@
+export default async function page ({req, props}) {
+  return JSON.stringify({one:1}, null, 2);
+
+
+}