+++ server/pages/api/two.json.js
@@ -1,0 +1,5 @@
+export default async function page ({req, props}) {
+  return JSON.stringify({two:2}, null, 2);
+
+
+}