+++ CHANGELOG.md
@@ -1,11 +1,12 @@
 
 # TODO
+- login
 - sqlite
 - colocated queries client/server sharing types
-- login
-- layout
-- 
+
 
 # DONE
+- layout
+- debug: lajout.js, on doubleclick reload page OR just css
 - added jsconfig.json for typecheck
 - moved to import